+++ src/moulti/widgets/__init__.py
@@ -3,4 +3,5 @@ WIDGET_MODULES = [
 	'step',
 	'buttonquestion',
 	'inputquestion',
+	'divider',
 ]